old_django_app/requirement_pip.txt
# pip install -r requirement_pip.txt
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
defusedxml==0.5.0
Django==1.11.5
django-cors-headers==2.1.0
django-filter==1.0.4
django-rest-swagger==2.1.2
djangorestframework==3.6.4
djangorestframework-jwt==1.11.0
factory-boy==2.10.0
Faker==0.8.12
flake8==3.4.1
idna==2.6
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
mccabe==0.6.1
oauthlib==2.0.3
openapi-codec==1.3.2
pep8==1.7.0
pycodestyle==2.3.1
pyflakes==1.5.0
PyJWT==1.5.3
python-dateutil==2.7.1
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
simplejson==3.11.1
six==1.10.0
social-auth-app-django==1.2.0
social-auth-core==1.4.0
text-unidecode==1.2
twitter==1.17.1
uritemplate==3.0.0
urllib3==1.22